Add instructions for raising proposals
