+++ Anchor.toml
@@ -22,6 +22,7 @@ autocrat_v0 = "yarn run ts-node scripts/autocratV0.ts"
 clob = "yarn run tsc scripts/clob.ts --esModuleInterop --resolveJsonModule && node scripts/clob.js && rm scripts/clob.js"
 crank = "yarn run ts-node scripts/crank.ts"
 main = "yarn run ts-node scripts/main.ts"
+propose = "yarn run ts-node scripts/initializeProposal.ts"
 initialize_vault = "yarn run tsc scripts/initializeVault.ts --esModuleInterop --resolveJsonModule && node scripts/initializeVault.js"
 place_order = "yarn run ts-node scripts/placeOrder.ts"
 test = "yarn run ts-mocha -p ./tsconfig.json -t 1000000 tests/*.ts"
