+++ README.md
@@ -5,6 +5,16 @@
 Monorepo that houses the Meta-DAO's programs. A description of these programs
 and what they do can be found at [themetadao.org](https://themetadao.org).
 
+## Scripts
+
+The scripts folder contains a few scripts that you can use to interact with the Meta-DAO.
+Today, the only way to create proposals is via script. You can do this by modifying the 
+`initializeProposal.ts` script and replacing its `pubkey`, `accounts`, and `data` with the
+SVM instruction that you want to use in your proposal.
+
+Then, run `anchor run propose --provider.cluster CLUSTER`, where `CLUSTER` is replaced with
+either devnet, mainnet, or (recommended) an RPC URL.
+
 ## Deployments
 
 | program           | tag | program ID                                  |
